Moved two pages and two hooks
