+++ src/hooks/nvi/useInstitutionReportsFilteredByUrl.ts
@@ -2,9 +2,9 @@ import { useSearchParams } from 'react-router';
 import { getNviInstitutionName } from '../../pages/basic_data/app_admin/nviAdmin/nviAdminHelpers';
 import { InstitutionReport } from '../../types/nvi.types';
 import { getDefaultNviYear } from '../../utils/hooks/useNviCandidatesParams';
-import { useFetchNviReports } from './useFetchNviReports';
+import { useFetchNviReports } from '../../api/hooks/useFetchNviReports';
 
-export const useGetUrlFilteredInstitutionReports = () => {
+export const useInstitutionReportsFilteredByUrl = () => {
   const [searchParams] = useSearchParams();
   const yearParam = searchParams.get('year')?.trim();
   const parsedYear = yearParam ? Number(yearParam) : NaN;
