+++ src/hooks/nvi/useInstitutionReportsSortedByUrl.ts
@@ -1,9 +1,10 @@
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
-import { getNviAdminSortValue, NviAdminOrderBy } from '../../pages/basic_data/app_admin/nviAdmin/nviAdminHelpers';
+import { NviAdminOrderBy } from '../../components/src/components/sortSelectors/sortNviTable/nviAdminSortHelpers';
+import { getNviAdminSortValue } from '../../pages/basic_data/app_admin/nviAdmin/nviAdminHelpers';
 import { InstitutionReport } from '../../types/nvi.types';
 
-export const useSortInstitutionReports = (reports: InstitutionReport[]) => {
+export const useInstitutionReportsSortedByUrl = (reports: InstitutionReport[]) => {
   const [searchParams] = useSearchParams();
   const { t } = useTranslation();
